expose build_ijar for scala_library (#1753)
Expose build_ijar to enable compiling scala code, that is not for macro
definitions, but would be used in inlining (so ijar support should be
disabled for it).

You may need strict dependency checks (transitive dependencies not on the
classpath) for code that is to be used in inlining.

Till now one would need to use scala_macro_library for that, but it was
recently switched to include all transitive dependencies on the compiler's
classpath, hence strict dependencies are not enforced there.

The new attribute enables having both:
a) code compiled for inlining
b) strict dependency checks